sql txt

+++ code/Parknshop/src/main/java/com/tmazon/service/impl/OrderServiceImpl.java
@@ -53,15 +53,17 @@ public class OrderServiceImpl implements OrderService {
 			}
 			orderInfo.setQuantity(cart.getQuantity());
 			orderInfo.setStatus(OrderInfo.STATUS_UNPAID);
-			if (orderInfoDao.insert(orderInfo)) {
-				int soldNum = product.getSoldNum() + 0;
-				product.setStockNum(product.getStockNum() - cart.getQuantity());
-				product.setSoldNum(soldNum + cart.getQuantity());
-				productDao.updateStockNum(product);
-				cartDao.delete(new Cart(order.getUserId(), orderInfo
-						.getProductId(), null));
-			} else {
-				flag = false;
+			if (orderInfo.getQuantity() != 0) {
+				if(orderInfoDao.insert(orderInfo)){
+					int soldNum = product.getSoldNum() + 0;
+					product.setStockNum(product.getStockNum() - cart.getQuantity());
+					product.setSoldNum(soldNum + cart.getQuantity());
+					productDao.updateStockNum(product);
+					cartDao.delete(new Cart(order.getUserId(), orderInfo
+							.getProductId(), null));
+				}else {
+					flag = false;
+				}
 			}
 		}
